requirements.txt
gpiozero
azure-iot-device
pigpio
lgpio
rpi-lgpio